fix the runtime warnings

+++ sarxarray/conf.py
@@ -1,4 +1,4 @@
 import numpy as np
 
 # Standard data types in sarxarray:
-_dtypes = dict(int=np.int32, float=np.float16, complex=np.complex64)
+_dtypes = dict(int=np.int32, float=np.float32, complex=np.complex64)
